feat: add constants dir
