+++ src/connect/constants/ui.tsx
@@ -1,0 +1,25 @@
+import { Icon } from '@/powerhouse';
+
+export const defaultDropdownMenuOptions = [
+    {
+        id: 'duplicate',
+        label: 'Duplicate',
+        icon: <Icon name="files-earmark" />,
+    },
+    {
+        id: 'new-folder',
+        label: 'New Folder',
+        icon: <Icon name="folder-plus" />,
+    },
+    {
+        id: 'rename',
+        label: 'Rename',
+        icon: <Icon name="pencil" />,
+    },
+    {
+        id: 'delete',
+        label: 'Delete',
+        icon: <Icon name="trash" />,
+        className: 'text-[#EA4335]',
+    },
+] as const;
