+++ src/connect/index.tsx
@@ -1,4 +1,5 @@
 export * from './components';
+export * from './constants';
 export * from './context';
 export * from './hooks';
 export * from './types';
